Fix float-to-int UB causing visual artifacts and crash on x86_64 Linux
The border width expressions `x1+16/zoom` and `x2-16/zoom` promote the
sentinel values 0x80000000/0x7fffffff to float before arithmetic.
(float)INT_MAX rounds up to 2147483648.0, and converting that back to
int is undefined behavior. On ARM64 (macOS) the FCVTZS instruction
saturates to INT_MAX so this went unnoticed; on x86_64 (Linux) CVTTSS2SI
wraps to INT_MIN, causing the entire right background to be drawn with
the 16px-wide border texture. Out-of-bounds texture reads produce the
visual artifacts, and in release builds hit unmapped memory (segfault).

Cast 16/zoom to int before the addition/subtraction so the sentinel
values stay in integer arithmetic and never get promoted to float.

Fixes #6

+++ src/renderframezoom.c
@@ -99,8 +99,8 @@ static Ptr DrawBorderZoomed(Ptr drawPos,int xDrawStart,int x1,int x2,int y,Ptr d
 
 static Ptr DrawBorderLineZoomed(Ptr drawPos,int xDrawStart,int x1,int x2,int y,Ptr data,Ptr leftBorder,Ptr rightBorder,float zoom)
 {
-	int leftBordEnd=x1+16/zoom;
-	int rightBordEnd=x2-16/zoom;
+	int leftBordEnd=x1+(int)(16/zoom);
+	int rightBordEnd=x2-(int)(16/zoom);
 	if(leftBordEnd>rightBordEnd)
 	{
 		leftBordEnd=x1+((x2-x1)>>1);
